move data and pointdata for exphist into data package
