+++ sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/data/DoubleExponentialHistogramData.java
@@ -3,19 +3,28 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.sdk.metrics.internal.aggregator;
+package io.opentelemetry.sdk.metrics.data;
 
 import com.google.auto.value.AutoValue;
-import io.opentelemetry.sdk.metrics.data.AggregationTemporality;
-import io.opentelemetry.sdk.metrics.data.ExponentialHistogramData;
-import io.opentelemetry.sdk.metrics.data.ExponentialHistogramPointData;
 import java.util.Collection;
 
+/**
+ * A simple, autovalue implementation of {@link ExponentialHistogramData}. For more detailed
+ * javadoc on the type, see {@link ExponentialHistogramData}.
+ *
+ * <p>See:
+ * https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/metrics/datamodel.md#exponentialhistogram
+ */
 @AutoValue
 public abstract class DoubleExponentialHistogramData implements ExponentialHistogramData {
 
   DoubleExponentialHistogramData() {}
 
+  /**
+   * Create a DoubleExponentialHistogramData.
+   *
+   * @return a DoubleExponentialHistogramData
+   */
   public static DoubleExponentialHistogramData create(
       AggregationTemporality temporality, Collection<ExponentialHistogramPointData> points) {
     return new AutoValue_DoubleExponentialHistogramData(temporality, points);
